requirements.txt
dkpro-cassis==0.5.1
spacy==3.0.5